Solved bugs
- Solved NullPointerException when a player that the plugin does not have to control quits the game
-  The plugin handles the case in which an afktime = 1 and AfkCheckInterval > 1.

+++ src/main/java/com/williamsimoni/afkfinder/QuitJoinEventHandler.java
@@ -63,6 +63,9 @@ public final class QuitJoinEventHandler implements Listener {
         Player player = e.getPlayer();
         PlayerData playerData = this.afkFinder.playersData.get(player.getUniqueId());
 
+        if (playerData == null)
+            return;
+
         if (this.afkFinder.afkZoneActive || !this.afkFinder.afkZoneTp){
             if (playerData.isAfk()) {
                 this.afkFinder.afkDatabase.removeAfkPlayerData(player);
